+++ backend/data/src/main/kotlin/io/tolgee/dtos/LLMProviderDto.kt
@@ -16,8 +16,8 @@ data class LLMProviderDto(
   override var deployment: String?,
   override var keepAlive: String?,
   override var format: String?,
-  override var pricePerMillionInput: Long?,
-  override var pricePerMillionOutput: Long?,
+  override var pricePerMillionInput: Double?,
+  override var pricePerMillionOutput: Double?,
 ) : LLMProviderInterface {
   @JsonSetter("type")
   fun setType(type: String) {
