+++ backend/data/src/main/kotlin/io/tolgee/model/LLMProvider.kt
@@ -24,8 +24,8 @@ class LLMProvider(
   override var deployment: String?,
   override var keepAlive: String?,
   override var format: String?,
-  override var pricePerMillionInput: Long?,
-  override var pricePerMillionOutput: Long?,
+  override var pricePerMillionInput: Double?,
+  override var pricePerMillionOutput: Double?,
   @ManyToOne
   @JoinColumn(name = "organization_id")
   var organization: Organization,
